// src/redux/users/types.ts
export enum LoadingStatuses {
  loading = 'LOADING',
  loaded = 'LOADED',
  error = 'ERROR',
}

export interface IUsersState {
  usersList: IUser[];
  hidedUsers: number[];
  usersLoadingStatus: LoadingStatuses;
  searchQuery: string;
  activeUser?: IUser;
}

export interface IUser {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string,
    suite: string,
    city: string,
    zipcode: string,
    geo: {
      lat: string,
      lng: string
    }
  };
  phone: string,
  website: string,
  company: {
    name: string,
    catchPhrase: string,
    bs: string
  }
}